+++ packages/app/src/flame/variations/parametric/post/symNetG12.tsx
@@ -45,7 +45,7 @@ export const symNetG12 = parametricVariation(
     const sy = P.spacey
     const px = pos.x + P.space
     const py = pos.y + P.space
-    const band = floor(random() * 8.0)
+    const band = f32(floor(random() * 8.0))
     if (band < 2.0) {
       if (band < 1.0) {
         return vec2f(px + sx, py + sy)
