+++ packages/app/src/flame/variations/parametric/post/symNetG5.tsx
@@ -33,7 +33,7 @@ export const symNetG5 = parametricVariation(
   SymNetG5ParamsEditor,
   (pos, _varInfo, P) => {
     'use gpu'
-    const band = floor(random() * 4.0)
+    const band = f32(floor(random() * 4.0))
     const hx = P.stepx * 0.5
     if (band < 1.0) {
       return vec2f(pos.x - P.sepx - 2.0 - hx, pos.y + P.sepy - 0.5)
